+++ FakeServer/Authentication/AddAuthorizationHeaderParameterOperationFilter.cs
@@ -1,0 +1,24 @@
+﻿using Swashbuckle.AspNetCore.Swagger;
+using Swashbuckle.AspNetCore.SwaggerGen;
+using System.Collections.Generic;
+
+namespace FakeServer.Authentication
+{
+    public class AddAuthorizationHeaderParameterOperationFilter : IOperationFilter
+    {
+        public void Apply(Operation operation, OperationFilterContext context)
+        {
+            if (operation.Parameters == null)
+                operation.Parameters = new List<IParameter>();
+
+            operation.Parameters.Add(new NonBodyParameter
+            {
+                Name = "Authorization",
+                In = "header",
+                Description = "Authorization header. Usage e.g.: Bearer [access_token]",
+                Required = false,
+                Type = "string"
+            });
+        }
+    }
+}