+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### [Unreleased]
 * Caching of unchanged resources with ETag and If-None-Match headers
 * Avoiding mid-air collisions with ETag and If-Match headers
+* Authorization header input to Swagger if authentication is enabled
 
 ### [0.5.0] - 2017-10-20
 * Support for HTTP HEAD Method